* sim_client.c (main): Cast argument to isupper()/tolower() to int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13377 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/simple/client/sim_client.c
@@ -165,8 +165,8 @@ main(argc, argv)
 
     /* lower-case to get name for "instance" part of service name */
     for (cp = full_hname; *cp; cp++)
-        if (isupper(*cp))
-            *cp = tolower(*cp);
+        if (isupper((int) *cp))
+            *cp = tolower((int) *cp);
 
     /* Set server's address */
     (void) memset((char *)&s_sock, 0, sizeof(s_sock));
